% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coerce.R
\name{as_cv_tobit}
\alias{as_cv_tobit}
\alias{as_cv_tobit.mean_tobit}
\alias{as_cv_tobit.geomean_tobit}
\title{Calculate the percent coefficient of variation (CV)}
\usage{
as_cv_tobit(x)

\method{as_cv_tobit}{mean_tobit}(x)

\method{as_cv_tobit}{geomean_tobit}(x)
}
\arguments{
\item{x}{The object to use for CV calculation}
}
\value{
The percent coefficient of variation as a numeric scalar
}
\description{
Calculate the percent coefficient of variation (CV)
}
\section{Methods (by class)}{
\itemize{
\item \code{mean_tobit}: CV for the arithmetic mean: \code{100*sd/mu}

\item \code{geomean_tobit}: CV for the geometric mean:
\code{100*sqrt(exp(sd^2)-1)}
}}

\seealso{
Other Class creation and coercion: 
\code{\link{as_geomean_tobit}()},
\code{\link{as_mean_tobit}()},
\code{\link{as_sd_tobit}()}
}
\concept{Class creation and coercion}
