% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coerce.R
\name{as_geomean_tobit}
\alias{as_geomean_tobit}
\alias{as_geomean_tobit.default}
\alias{as_geomean_tobit.mean_tobit}
\title{Create or cast to a geomean_tobit object}
\usage{
as_geomean_tobit(mu, sd, message = NULL)

\method{as_geomean_tobit}{default}(mu, sd, message = NULL)

\method{as_geomean_tobit}{mean_tobit}(mu, sd, message = NULL)
}
\arguments{
\item{mu}{The geometric mean}

\item{sd}{The geometric standard deviation}

\item{message}{Additional information about the mean}
}
\value{
A 'geomean_tobit' class object which is a number with attributes for
'sd' and 'message'.
}
\description{
Create or cast to a geomean_tobit object
}
\examples{
as_geomean_tobit(mu=10, sd=2, message="This is just an example")
}
\seealso{
Other Class creation and coercion: 
\code{\link{as_cv_tobit}()},
\code{\link{as_mean_tobit}()},
\code{\link{as_sd_tobit}()}
}
\concept{Class creation and coercion}
