% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coerce.R
\name{as_mean_tobit}
\alias{as_mean_tobit}
\alias{as_mean_tobit.default}
\title{Create or cast to a mean_tobit object}
\usage{
as_mean_tobit(mu, sd, message = NULL)

\method{as_mean_tobit}{default}(mu, sd, message = NULL)
}
\arguments{
\item{mu}{The arithmetic mean}

\item{sd}{The arithmetic standard deviation}

\item{message}{Additional information about the mean}
}
\value{
A 'mean_tobit' class object which is a number with attributes for
'sd' and 'message'.
}
\description{
Create or cast to a mean_tobit object
}
\examples{
as_mean_tobit(mu=10, sd=2, message="This is just an example")
}
\seealso{
Other Class creation and coercion: 
\code{\link{as_cv_tobit}()},
\code{\link{as_geomean_tobit}()},
\code{\link{as_sd_tobit}()}
}
\concept{Class creation and coercion}
