% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coerce.R
\name{as_sd_tobit}
\alias{as_sd_tobit}
\alias{as_sd_tobit.default}
\alias{as_sd_tobit.mean_tobit}
\alias{as_sd_tobit.geomean_tobit}
\title{Create or extract the standard deviation from a mean_tobit object}
\usage{
as_sd_tobit(mu, sd, message = NULL)

\method{as_sd_tobit}{default}(mu, sd, message = NULL)

\method{as_sd_tobit}{mean_tobit}(mu, sd, message = NULL)

\method{as_sd_tobit}{geomean_tobit}(mu, sd, message = NULL)
}
\arguments{
\item{mu}{The arithmetic mean}

\item{sd}{The arithmetic standard deviation}

\item{message}{Additional information about the mean}
}
\value{
A 'sd_tobit' class object which is a number with attributes of 'mu'
(the mean) and 'message'.
}
\description{
Create or extract the standard deviation from a mean_tobit object
}
\section{Methods (by class)}{
\itemize{
\item \code{mean_tobit}: Convert a 'mean_tobit' object to a 'sd_tobit' object

\item \code{geomean_tobit}: Convert a 'mean_tobit' object to a 'sd_tobit' object
}}

\seealso{
Other Class creation and coercion: 
\code{\link{as_cv_tobit}()},
\code{\link{as_geomean_tobit}()},
\code{\link{as_mean_tobit}()}
}
\concept{Class creation and coercion}
