% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_geomean.R
\name{geomean_tobit}
\alias{geomean_tobit}
\title{Calculate the censored geometric mean}
\usage{
geomean_tobit(x, lower = 0, upper = Inf, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{lower, upper}{The lower and upper bounds for censoring \code{x}}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped from \code{x} before the computation proceeds.}

\item{...}{
  Arguments passed on to \code{\link[=mean_tobit]{mean_tobit}}
  \describe{
    \item{\code{lower,upper}}{The lower and upper bounds for censoring \code{x}}
    \item{\code{include_equal,include_equal_lower,include_equal_upper}}{Should the
bounds consider \code{x <= lower} and \code{x >= upper} (if \code{TRUE}) or
\code{x < lower} and \code{x > upper} (if \code{FALSE}) to be censored?
(\code{include_equal} is overridden by setting \code{include_equal_lower}
or \code{include_equal_upper}.)}
    \item{\code{distribution}}{What distribution (Student "t" or "normal") should be
used.  If "t", the degrees of freedom are the number of non-\code{NA}
values in \code{x}.}
  }}
}
\value{
A "sd_tobit" class object
}
\description{
Calculate the censored geometric mean
}
\seealso{
Other Censored Descriptive Statistics: 
\code{\link{mean_tobit}()},
\code{\link{sd_tobit}()}
}
\concept{Censored Descriptive Statistics}
