% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_geomean.R
\name{mean_tobit}
\alias{mean_tobit}
\title{Calculate a censored mean}
\usage{
mean_tobit(
  x,
  lower = -Inf,
  upper = Inf,
  include_equal = TRUE,
  include_equal_lower = include_equal,
  include_equal_upper = include_equal,
  distribution,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric vector}

\item{lower, upper}{The lower and upper bounds for censoring \code{x}}

\item{include_equal, include_equal_lower, include_equal_upper}{Should the
bounds consider \code{x <= lower} and \code{x >= upper} (if \code{TRUE}) or
\code{x < lower} and \code{x > upper} (if \code{FALSE}) to be censored?
(\code{include_equal} is overridden by setting \code{include_equal_lower}
or \code{include_equal_upper}.)}

\item{distribution}{What distribution (Student "t" or "normal") should be
used.  If "t", the degrees of freedom are the number of non-\code{NA}
values in \code{x}.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped from \code{x} before the computation proceeds.}

\item{...}{
  Arguments passed on to \code{\link[stats:optim]{stats::optim}}
  \describe{
    \item{\code{par}}{Initial values for the parameters to be optimized over.}
    \item{\code{fn}}{A function to be minimized (or maximized), with first
   argument the vector of parameters over which minimization is to take
   place.  It should return a scalar result.}
    \item{\code{gr}}{A function to return the gradient for the \code{"BFGS"},
   \code{"CG"} and \code{"L-BFGS-B"} methods.  If it is \code{NULL}, a
   finite-difference approximation will be used.

   For the \code{"SANN"} method it specifies a function to generate a new
   candidate point.  If it is \code{NULL} a default Gaussian Markov
   kernel is used.}
    \item{\code{method}}{The method to be used. See \sQuote{Details}.  Can be abbreviated.}
    \item{\code{lower,upper}}{Bounds on the variables for the \code{"L-BFGS-B"}
   method, or bounds in which to \emph{search} for method \code{"Brent"}.}
    \item{\code{control}}{a \code{\link{list}} of control parameters.  See \sQuote{Details}.}
    \item{\code{hessian}}{Logical. Should a numerically differentiated Hessian
   matrix be returned?}
  }}
}
\value{
A 'mean_tobit' object
}
\description{
Calculate a censored mean
}
\seealso{
Other Censored Descriptive Statistics: 
\code{\link{geomean_tobit}()},
\code{\link{sd_tobit}()}
}
\concept{Censored Descriptive Statistics}
