% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{negLogLik_tobit}
\alias{negLogLik_tobit}
\title{Estimate the log-likelihood for the gamma/delta transformed convex parameters
for tobit regression}
\usage{
negLogLik_tobit(
  param,
  x,
  lower_limit,
  upper_limit,
  mask_lower,
  mask_between,
  mask_upper,
  distribution = c("t", "norm")
)
}
\arguments{
\item{param}{c(gamma, delta) see details}

\item{x}{The vector of values}

\item{lower_limit, upper_limit}{The vector lower and upper limits for censoring}

\item{mask_lower, mask_between, mask_upper}{Boolean masks indicating if the
value is below the lower limit, a point value, or above the upper limit}

\item{distribution}{The distribution to use}
}
\description{
Estimate the log-likelihood for the gamma/delta transformed convex parameters
for tobit regression
}
\details{
\code{beta = delta/gamma} and \code{sigma^2 = gamma^-2}.  The estimation is based on
}
\references{
Olsen RJ. Note on the Uniqueness of the Maximum Likelihood
Estimator for the Tobit Model. Econometrica. 1978;46(5):1211.
doi:10.2307/1911445
}
\keyword{Internal}
