% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_geomean.R
\name{sd_tobit}
\alias{sd_tobit}
\title{Calculate the censored standard deviation}
\usage{
sd_tobit(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=mean_tobit]{mean_tobit}}
  \describe{
    \item{\code{x}}{A numeric vector}
    \item{\code{lower,upper}}{The lower and upper bounds for censoring \code{x}}
    \item{\code{include_equal,include_equal_lower,include_equal_upper}}{Should the
bounds consider \code{x <= lower} and \code{x >= upper} (if \code{TRUE}) or
\code{x < lower} and \code{x > upper} (if \code{FALSE}) to be censored?
(\code{include_equal} is overridden by setting \code{include_equal_lower}
or \code{include_equal_upper}.)}
    \item{\code{distribution}}{What distribution (Student "t" or "normal") should be
used.  If "t", the degrees of freedom are the number of non-\code{NA}
values in \code{x}.}
    \item{\code{na.rm}}{A logical value indicating whether \code{NA} values should be
stripped from \code{x} before the computation proceeds.}
  }}
}
\value{
A "sd_tobit" class object
}
\description{
Since the standard deviation is almost always desired with the mean, it is
more efficient to use \code{as_sd_tobit()} on the mean object instead of
calculating the standard deviation separately.
}
\seealso{
Other Censored Descriptive Statistics: 
\code{\link{geomean_tobit}()},
\code{\link{mean_tobit}()}
}
\concept{Censored Descriptive Statistics}
